# contracts/openapi/hive-core-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HIVE Core API",
    "summary": "Population-level security analysis for autonomous agent estates.",
    "description": "HIVE compares a declared architecture against observed agent interactions, reports capability compositions the architecture forbids, and evaluates pre-authorised containment actions against the live graph.\n\nThis service runs entirely locally against fictional fixtures. It contacts no external system and all controls are simulated.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "service"
        ],
        "summary": "Liveness and operating mode",
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Api V1 Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scenarios": {
      "get": {
        "tags": [
          "replay"
        ],
        "summary": "List available scenarios",
        "operationId": "list_scenarios_api_v1_scenarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Scenarios Api V1 Scenarios Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replays/{scenario_id}/reset": {
      "post": {
        "tags": [
          "replay"
        ],
        "summary": "Restart a replay",
        "operationId": "reset_replay_api_v1_replays__scenario_id__reset_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reset Replay Api V1 Replays  Scenario Id  Reset Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replays/{scenario_id}/advance": {
      "post": {
        "tags": [
          "replay"
        ],
        "summary": "Advance a replay",
        "operationId": "advance_replay_api_v1_replays__scenario_id__advance_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          },
          {
            "name": "to_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Advance to this sequence number. Omit to step one event.",
              "title": "To Sequence"
            },
            "description": "Advance to this sequence number. Omit to step one event."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Advance Replay Api V1 Replays  Scenario Id  Advance Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replays/{scenario_id}/run": {
      "post": {
        "tags": [
          "replay"
        ],
        "summary": "Run a replay to the end",
        "operationId": "run_replay_api_v1_replays__scenario_id__run_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Run Replay Api V1 Replays  Scenario Id  Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replays/{scenario_id}/state": {
      "get": {
        "tags": [
          "replay"
        ],
        "summary": "Current replay state",
        "operationId": "get_state_api_v1_replays__scenario_id__state_get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get State Api V1 Replays  Scenario Id  State Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replays/{scenario_id}/findings": {
      "get": {
        "tags": [
          "detection"
        ],
        "summary": "Findings and containment plans for the current state",
        "operationId": "get_findings_api_v1_replays__scenario_id__findings_get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Findings Api V1 Replays  Scenario Id  Findings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replays/{scenario_id}/findings/{finding_id}": {
      "get": {
        "tags": [
          "detection"
        ],
        "summary": "One finding with its plan",
        "operationId": "get_finding_api_v1_replays__scenario_id__findings__finding_id__get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          },
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Finding Api V1 Replays  Scenario Id  Findings  Finding Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replays/{scenario_id}/plans/{plan_id}/apply": {
      "post": {
        "tags": [
          "containment"
        ],
        "summary": "Issue the recommended simulated control and verify it",
        "operationId": "apply_plan_api_v1_replays__scenario_id__plans__plan_id__apply_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          },
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Apply Plan Api V1 Replays  Scenario Id  Plans  Plan Id  Apply Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replays/{scenario_id}/architecture": {
      "get": {
        "tags": [
          "architecture"
        ],
        "summary": "The declared architecture this scenario is measured against",
        "operationId": "get_architecture_api_v1_replays__scenario_id__architecture_get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Architecture Api V1 Replays  Scenario Id  Architecture Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replays/{scenario_id}/immunity": {
      "get": {
        "tags": [
          "immunity"
        ],
        "summary": "Reviewed patterns recorded from contained findings",
        "operationId": "list_immunity_api_v1_replays__scenario_id__immunity_get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Immunity Api V1 Replays  Scenario Id  Immunity Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replays/{scenario_id}/immunity/{pattern_id}/promote": {
      "post": {
        "tags": [
          "immunity"
        ],
        "summary": "Advance a pattern through its review lifecycle",
        "operationId": "promote_immunity_api_v1_replays__scenario_id__immunity__pattern_id__promote_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          },
          {
            "name": "pattern_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pattern Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Promote Immunity Api V1 Replays  Scenario Id  Immunity  Pattern Id  Promote Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lab/simulate": {
      "post": {
        "tags": [
          "lab"
        ],
        "summary": "Synthesise a population and run the real detection pipeline on it",
        "operationId": "simulate_swarm_api_v1_lab_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwarmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Simulate Swarm Api V1 Lab Simulate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lab/options": {
      "get": {
        "tags": [
          "lab"
        ],
        "summary": "Valid lab parameter values",
        "operationId": "lab_options_api_v1_lab_options_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Lab Options Api V1 Lab Options Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PromoteRequest": {
        "properties": {
          "to": {
            "type": "string",
            "title": "To",
            "description": "Target lifecycle stage: shadow, active, or draft."
          }
        },
        "type": "object",
        "required": [
          "to"
        ],
        "title": "PromoteRequest"
      },
      "SwarmRequest": {
        "properties": {
          "population": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 3.0,
            "title": "Population",
            "default": 24
          },
          "connectivity": {
            "type": "string",
            "title": "Connectivity",
            "default": "normal"
          },
          "shared_memory": {
            "type": "string",
            "title": "Shared Memory",
            "default": "limited"
          },
          "delegation": {
            "type": "string",
            "title": "Delegation",
            "default": "normal"
          },
          "external_access": {
            "type": "string",
            "title": "External Access",
            "default": "limited"
          },
          "perturbation": {
            "type": "string",
            "title": "Perturbation",
            "default": "unregistered_shared_resource"
          }
        },
        "type": "object",
        "title": "SwarmRequest",
        "description": "Population parameters for one lab experiment."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development server"
    }
  ]
}